feat: Extract email template into separate component
